+++ CHANGES.md
@@ -1,3 +1,10 @@
-0.0.1
+0.1.0
 -----
-Skeleton
+Basic functions implemented:
+- Support package definition, file name is lnkr-export.toml
+  - can export local folder/files
+  - can wrap other projects with wrapper definition (file name is lnkr-wrapper.toml)
+- Support import definition, file name is lnkr-import.toml
+  - support folder, and files
+  - support copy mode and symlink mode
+- Doc is missing for now